+++ sdk/ai/Azure.AI.Extensions.OpenAI/tests/Samples/Sample_WebSearch.cs
@@ -9,6 +9,7 @@ using NUnit.Framework;
 using OpenAI.Responses;
 using Azure.AI.Projects;
 using Azure.AI.Projects.Agents;
+using System.Text;
 
 namespace Azure.AI.Extensions.OpenAI.Tests.Samples;
 
@@ -21,17 +22,17 @@ public class Sample_WebSearch : ProjectsOpenAITestBase
         IgnoreSampleMayBe();
         #region Snippet:Sample_CreateAgentClient_WebSearch
 #if SNIPPET
-        var projectEndpoint = System.Environment.GetEnvironmentVariable("PROJECT_ENDPOINT");
-        var modelDeploymentName = System.Environment.GetEnvironmentVariable("MODEL_DEPLOYMENT_NAME");
+        var projectEndpoint = System.Environment.GetEnvironmentVariable("FOUNDRY_PROJECT_ENDPOINT");
+        var modelDeploymentName = System.Environment.GetEnvironmentVariable("FOUNDRY_MODEL_NAME");
 #else
-        var projectEndpoint = TestEnvironment.PROJECT_ENDPOINT;
-        var modelDeploymentName = TestEnvironment.MODELDEPLOYMENTNAME;
+        var projectEndpoint = TestEnvironment.FOUNDRY_PROJECT_ENDPOINT;
+        var modelDeploymentName = TestEnvironment.FOUNDRY_MODEL_NAME;
 #endif
         AIProjectClient projectClient = new(endpoint: new Uri(projectEndpoint), tokenProvider: new DefaultAzureCredential());
 
         #endregion
         #region Snippet:Sample_CreateAgent_WebSearch_Async
-        PromptAgentDefinition agentDefinition = new(model: modelDeploymentName)
+        DeclarativeAgentDefinition agentDefinition = new(model: modelDeploymentName)
         {
             Instructions = "You are a helpful assistant that can search the web",
             Tools = { ResponseTool.CreateWebSearchTool(userLocation: WebSearchToolLocation.CreateApproximateLocation(country: "GB", city: "London", region: "London")), }
@@ -48,7 +49,7 @@ public class Sample_WebSearch : ProjectsOpenAITestBase
 
         #region Snippet:Sample_WaitForResponse_WebSearch
         Assert.That(response.Status, Is.EqualTo(ResponseStatus.Completed));
-        Console.WriteLine(response.GetOutputText());
+        Console.WriteLine($"{response.GetOutputText()} {GetFormattedAnnotation(response)}");
         #endregion
 
         #region Snippet:Sample_Cleanup_WebSearch_Async
@@ -62,16 +63,16 @@ public class Sample_WebSearch : ProjectsOpenAITestBase
     {
         IgnoreSampleMayBe();
 #if SNIPPET
-        var projectEndpoint = System.Environment.GetEnvironmentVariable("PROJECT_ENDPOINT");
-        var modelDeploymentName = System.Environment.GetEnvironmentVariable("MODEL_DEPLOYMENT_NAME");
+        var projectEndpoint = System.Environment.GetEnvironmentVariable("FOUNDRY_PROJECT_ENDPOINT");
+        var modelDeploymentName = System.Environment.GetEnvironmentVariable("FOUNDRY_MODEL_NAME");
 #else
-        var projectEndpoint = TestEnvironment.PROJECT_ENDPOINT;
-        var modelDeploymentName = TestEnvironment.MODELDEPLOYMENTNAME;
+        var projectEndpoint = TestEnvironment.FOUNDRY_PROJECT_ENDPOINT;
+        var modelDeploymentName = TestEnvironment.FOUNDRY_MODEL_NAME;
 #endif
         AIProjectClient projectClient = new(endpoint: new Uri(projectEndpoint), tokenProvider: new DefaultAzureCredential());
 
         #region Snippet:Sample_CreateAgent_WebSearch_Sync
-        PromptAgentDefinition agentDefinition = new(model: modelDeploymentName)
+        DeclarativeAgentDefinition agentDefinition = new(model: modelDeploymentName)
         {
             Instructions = "You are a helpful assistant that can search the web",
             Tools = { ResponseTool.CreateWebSearchTool(userLocation: WebSearchToolLocation.CreateApproximateLocation(country: "GB", city: "London", region: "London")), }
@@ -87,12 +88,41 @@ public class Sample_WebSearch : ProjectsOpenAITestBase
         #endregion
 
         Assert.That(response.Status, Is.EqualTo(ResponseStatus.Completed));
-        Console.WriteLine(response.GetOutputText());
+        Console.WriteLine($"{response.GetOutputText()} {GetFormattedAnnotation(response)}");
 
         #region Snippet:Sample_Cleanup_WebSearch_Sync
         projectClient.Agents.DeleteAgentVersion(agentName: agentVersion.Name, agentVersion: agentVersion.Version);
         #endregion
     }
+
+    #region Snippet:Sample_FormatReference_WebSearch
+    private static string GetFormattedAnnotation(ResponseResult results)
+    {
+        StringBuilder references = new();
+        foreach (ResponseItem item in results.OutputItems)
+        {
+            if (item is MessageResponseItem messageItem)
+            {
+                foreach (ResponseContentPart content in messageItem.Content)
+                {
+                    foreach (ResponseMessageAnnotation annotation in content.OutputTextAnnotations)
+                    {
+                        if (annotation is UriCitationMessageAnnotation uriAnnotation)
+                        {
+                            references.Append($"[{uriAnnotation.Title}]({uriAnnotation.Uri}),");
+                        }
+                    }
+                }
+            }
+        }
+        if (references.Length > 0)
+        {
+            // Remove the last comma.
+            references.Remove(references.Length - 1, 1);
+        }
+        return references.ToString();
+    }
+    #endregion
 
     public Sample_WebSearch(bool isAsync) : base(isAsync)
     { }