+++ sdk/dataprotection/Azure.ResourceManager.DataProtectionBackup/src/Customized/Models/DataProtectionBackupVaultProperties.cs
@@ -1,18 +1,37 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-using System;
-using System.Collections.Generic;
+
 using System.ComponentModel;
 
 #nullable disable
 
+// NOTE: The following customization is intentionally retained for backward compatibility.
 namespace Azure.ResourceManager.DataProtectionBackup.Models
 {
     /// <summary> Backup Vault. </summary>
     public partial class DataProtectionBackupVaultProperties
     {
         /// <summary> Is vault protected by resource guard. </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public bool? IsVaultProtectedByResourceGuard { get; set; }
+
+        /// <summary> Gets or sets the alert settings for all job failures. </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public AzureMonitorAlertsState? AlertSettingsForAllJobFailures
+        {
+            get
+            {
+                return MonitoringSettings is null ? default : MonitoringSettings.AlertSettingsForAllJobFailures;
+            }
+            set
+            {
+                if (MonitoringSettings is null)
+                {
+                    MonitoringSettings = new MonitoringSettings();
+                }
+                MonitoringSettings.AlertSettingsForAllJobFailures = value;
+            }
+        }
 
         /// <summary>
         /// Gets or sets the cross subscription restore state.