+++ sdk/provisioning/Generator/src/Model/Specification.Files.cs
@@ -15,7 +15,9 @@ public abstract partial class Specification : ModelBase
 
         // TODO: This assumes we're always running in place, in the repo
         string? path = Environment.CurrentDirectory;
-        while (path is not null && !Directory.Exists(Path.Combine(path, ".git")))
+        while (path is not null &&
+               !Directory.Exists(Path.Combine(path, ".git")) &&
+               !File.Exists(Path.Combine(path, ".git")))
         {
             // Walk up a level
             path = Path.GetDirectoryName(path);
@@ -24,8 +26,8 @@ public abstract partial class Specification : ModelBase
         // If all else fails, just use the current directory
         path ??= Environment.CurrentDirectory;
 
-        // Walk from the root of the repo into the provisioning folder
-        path = Path.Combine(path, "sdk", "provisioning");
+        // Walk from the root of the repo into the service folder
+        path = Path.Combine(path, "sdk", ServiceDirectory);
         if (!Directory.Exists(path))
         {
             throw new InvalidOperationException($"Directory {path} must exist to write {Namespace}!");
@@ -189,17 +191,6 @@ public abstract partial class Specification : ModelBase
 
                 </Project>
                 """);
-        Directory.CreateDirectory(Path.Combine(path, "src", "Properties"));
-        File.WriteAllText(Path.Combine(path, "src", "Properties", "AssemblyInfo.cs"),
-            """
-                // Copyright (c) Microsoft Corporation. All rights reserved.
-                // Licensed under the MIT License.
-
-                using System.Diagnostics.CodeAnalysis;
-
-                [assembly: Experimental("AZPROVISION001")]
-                """);
-
         // Generate the tests
         Directory.CreateDirectory(Path.Combine(path, "tests"));
         File.WriteAllText(Path.Combine(path, "tests", $"{Namespace}.Tests.csproj"),